requirements.txt
discord.py
asyncio
Status
PyNaCl
config
youtube_dl
environ
lxml
asyncio
datetime
requests
cleverwrap
utils
aiohttp
websockets
BeautifulSoup4
bs4
datetime



